pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel>=0.41.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lesson-generator"
version = "0.1.0"
description = "AI-powered lesson generator for programming courses"
authors = [
    {name = "LibertyQuinzel", email = "your-email@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Education",
    "Topic :: Software Development :: Code Generators",
]
keywords = ["lesson", "generator", "ai", "education", "programming", "openai"]

dependencies = [
    "click>=8.1.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "pathlib2>=2.3.7",
    "typing-extensions>=4.7.0",
    # Web interface dependencies
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "websockets>=12.0",
    # Database dependencies
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "pylint>=2.17.0", 
    "mypy>=1.5.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-asyncio>=0.21.0",
    "pre-commit>=3.3.0",
]

[project.urls]
Homepage = "https://github.com/LibertyQuinzel/lesson-generator"
Repository = "https://github.com/LibertyQuinzel/lesson-generator"
Documentation = "https://github.com/LibertyQuinzel/lesson-generator#readme"
"Bug Tracker" = "https://github.com/LibertyQuinzel/lesson-generator/issues"

[project.scripts]
lesson-generator = "lesson_generator.cli:main"
lesson-generator-web = "lesson_generator.web.main:run_server"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
lesson_generator = ["templates/*.j2", "templates/**/*.j2"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pylint.messages_control]
disable = [
    "C0114",  # missing-module-docstring
    "C0116",  # missing-function-docstring
    "R0903",  # too-few-public-methods
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src/lesson_generator",
    "--cov-report=term-missing",
    "--cov-report=html:coverage_html",
    "--cov-fail-under=90"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "e2e: End-to-end tests",
    "slow: Slow running tests",
]